Rename Heartcheck::Services::Formatters::* -> Heartcheck::Formatters::*
