+++ lib/heartcheck.rb
@@ -6,6 +6,8 @@ module Heartcheck
   require 'heartcheck/errors'
   require 'heartcheck/services'
   require 'heartcheck/logger'
+  require 'heartcheck/formatters/base'
+  require 'heartcheck/formatters/hash_response'
 
   @checks = []
 
@@ -16,7 +18,7 @@ module Heartcheck
     # @attr [Heartcheck::Executors::Base] the checks executor backend
     attr_accessor :executor
 
-    # @attr [Heartcheck::Executors::Base] type of response format
+    # @attr [Heartcheck::Formatters::Base] formats heart check result to something JSON encodedable
     attr_accessor :formatter
 
     # @attr_writer [Object] change the default logger
@@ -98,18 +100,18 @@ module Heartcheck
       checks.select { |ctx| ctx.respond_to?(:info) }
     end
 
-    # an formatter class that respond to format(checkers)
+    # transforms check results to a response
     #
-    # @return [Heartcheck::Services::Formatters::Base] or nil
+    # @return [Heartcheck::Formatters::Base] or nil
     def formatter
-      @formatter ||= Heartcheck::Services::Formatters::Base.new
+      @formatter ||= Heartcheck::Formatters::Base.new
     end
 
     # change current formatter to a hash response
     #
-    # @return [Heartcheck::Services::Formatters::HashResponse]
+    # @return [Heartcheck::Formatters::HashResponse]
     def use_hash_formatter!
-      self.formatter = Heartcheck::Services::Formatters::HashResponse.new
+      self.formatter = Heartcheck::Formatters::HashResponse.new
     end
 
     # an executor class that respond to dispatch(checkers)
