+++ lib/heartcheck/formatters/base.rb
@@ -1,0 +1,9 @@
+module Heartcheck
+  module Formatters
+    class Base
+      def format(checks)
+        checks
+      end
+    end
+  end
+end
