+++ lib/heartcheck/services.rb
@@ -1,3 +1,1 @@
 require 'heartcheck/services/firewall'
-require 'heartcheck/services/formatters/base'
-require 'heartcheck/services/formatters/hash_response'
