--- lib/heartcheck/services/formatters/base.rb
@@ -1,11 +1,0 @@
-module Heartcheck
-  module Services
-    module Formatters    
-      class Base 
-        def format(checks)  
-          checks
-        end
-      end
-    end
-  end
-end
