+++ spec/lib/heartcheck/formatters/hash_response_spec.rb
@@ -1,8 +1,9 @@
 require 'spec_helper'
 
-describe Heartcheck::Services::Formatters::Base do
+describe Heartcheck::Formatters::HashResponse do
   subject do
     Heartcheck.setup do |monitor|
+      monitor.use_hash_formatter!
     end
     Heartcheck.formatter
   end
@@ -12,8 +13,8 @@ describe Heartcheck::Services::Formatters::Base do
 
     it "when hash_formatter is true" do
       expect(subject.format([check_01,check_02])).to be_eql(
-          [{:dummy1=>{:status=>:ok}, :time=>1100},
-            {:dummy2=>{:status=>:ok}, :time=>100}])
+        {:dummy1=>{:status=>:ok, :time=>1100},
+         :dummy2=>{:status=>:ok, :time=>100}})
     end
   end
